project/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
numpy>=1.23.0
Pillow>=9.0.0
torch>=2.0.0
